feat: Add support for minimum number of floating IPs in the pool
This commit adds a new field, `MinIPs`, to the `OpenStackFloatingIPPoolSpec` struct. The `MinIPs` field represents the minimum number of floating IPs that should be available in the pool. If this field is set, the pool will start allocating floating IPs and will stop when it reaches this number of claimed IPs. If the `MinIPs` field is not set, there is no limit on the number of floating IPs in the pool.

The purpose of this change is to provide more flexibility and control over the number of floating IPs in the pool.

Refactor the code in `conditions_consts.go` to include a new constant, `MinIPsNotReachedReason`, which will be set when the minimum number of floating IPs has not been reached.

Refactor the code in `openstackfloatingippool_types.go` to include the new `MinIPs` field in the `OpenStackFloatingIPPoolSpec` struct.

This commit addresses the need to support a minimum number of floating IPs in the pool and improves the overall code structure and organization.